mss_evaluation/oracle-cqt/requirements.txt
scipy
numpy
musdb
cupy-cuda112
git+git://github.com/sevagh/sigsep-mus-eval@feat/cupy-accel#egg=museval
nsgt
pandas
seaborn
line_profiler
memory_profiler
tqdm
